+++ packages/colony-js-contract-client/src/__tests__/ContractMethodSender.js
@@ -56,6 +56,7 @@ describe('ContractMethodSender', () => {
   };
   const adapter = {
     getTransactionReceipt: sandbox.fn(async () => Promise.resolve(receipt)),
+    waitForTransaction: sandbox.fn(async () => Promise.resolve(transaction)),
   };
   const client = new ContractClient({ contract, adapter });
   sandbox
@@ -155,6 +156,9 @@ describe('ContractMethodSender', () => {
       functionName,
       eventHandlers,
     });
+    sandbox
+      .spyOn(method, '_waitForTransactionReceipt')
+      .mockImplementation(async () => receipt);
 
     const response = await method._sendWithWaitingForMining(
       transaction,
@@ -171,9 +175,6 @@ describe('ContractMethodSender', () => {
     });
 
     // Event data, transaction receipts and success are collected
-    expect(method.client.adapter.getTransactionReceipt).toHaveBeenCalledWith(
-      transaction.hash,
-    );
     expect(method.client.getEventData).toHaveBeenCalledWith({
       events: method.eventHandlers,
       timeoutMs: options.timeoutMs,
@@ -185,7 +186,7 @@ describe('ContractMethodSender', () => {
       timeoutMs: options.timeoutMs,
       transactionHash: transaction.hash,
     });
-    expect(method.client.adapter.getTransactionReceipt).toHaveBeenCalledWith(
+    expect(method._waitForTransactionReceipt).toHaveBeenCalledWith(
       transaction.hash,
     );
   });
@@ -197,6 +198,9 @@ describe('ContractMethodSender', () => {
       functionName,
       eventHandlers,
     });
+    sandbox
+      .spyOn(method, '_waitForTransactionReceipt')
+      .mockImplementation(async () => receipt);
 
     const response = method._sendWithoutWaitingForMining(
       transaction,
@@ -219,9 +223,26 @@ describe('ContractMethodSender', () => {
       timeoutMs: options.timeoutMs,
       transactionHash: transaction.hash,
     });
-    expect(method.client.adapter.getTransactionReceipt).toHaveBeenCalledWith(
+    expect(method._waitForTransactionReceipt).toHaveBeenCalledWith(
       transaction.hash,
     );
+  });
+
+  test('Waiting for transaction receipt', async () => {
+    const method = new ContractMethodSender({
+      client,
+      input,
+      functionName,
+      eventHandlers,
+    });
+    const hash = 'the tx hash';
+    const txReceipt = await method._waitForTransactionReceipt(hash);
+
+    expect(txReceipt).toEqual(receipt);
+    expect(method.client.adapter.getTransactionReceipt).toHaveBeenCalledWith(
+      hash,
+    );
+    expect(method.client.adapter.waitForTransaction).toHaveBeenCalledWith(hash);
   });
 
   test('Default send options', () => {