+++ packages/colony-js-contract-client/src/classes/ContractMethodSender.js
@@ -66,36 +66,22 @@ export default class ContractMethodSender<
     transaction: Transaction,
     timeoutMs: number,
   ): Promise<ContractResponse<OutputValues>> {
-    const receipt = await raceAgainstTimeout(
-      this.client.adapter.getTransactionReceipt(transaction.hash),
-      timeoutMs,
-    );
-
-    const meta = {
-      transaction,
-      receipt,
-    };
-    const successful = receipt.status === 1;
-
-    // If the receipt wasn't successful, return immediately rather than waiting
-    // for events/mined tx
-    if (!successful) {
-      return {
-        successful,
-        meta,
-        eventData: {},
-      };
-    }
-
     const eventData = await this.client.getEventData({
       events: this.eventHandlers,
       timeoutMs,
       transactionHash: transaction.hash,
     });
+    const receipt = await raceAgainstTimeout(
+      this._waitForTransactionReceipt(transaction.hash),
+      timeoutMs,
+    );
 
     return {
-      successful,
-      meta,
+      successful: receipt && receipt.status === 1,
+      meta: {
+        transaction,
+        receipt,
+      },
       eventData,
     };
   }
@@ -104,30 +90,30 @@ export default class ContractMethodSender<
     transaction: Transaction,
     timeoutMs: number,
   ): ContractResponse<OutputValues> {
+    const eventDataPromise = this.client.getEventData({
+      events: this.eventHandlers,
+      timeoutMs,
+      transactionHash: transaction.hash,
+    });
     const receiptPromise = raceAgainstTimeout(
-      this.client.adapter.getTransactionReceipt(transaction.hash),
+      this._waitForTransactionReceipt(transaction.hash),
       timeoutMs,
     );
     const successfulPromise = new Promise(async (resolve, reject) => {
       try {
-        const { status } = await receiptPromise;
-        resolve(status === 1);
+        const receipt = await receiptPromise;
+        resolve(receipt && receipt.status === 1);
       } catch (error) {
         reject(error.toString());
       }
     });
-
     return {
       successfulPromise,
       meta: {
         receiptPromise,
         transaction,
       },
-      eventDataPromise: this.client.getEventData({
-        events: this.eventHandlers,
-        timeoutMs,
-        transactionHash: transaction.hash,
-      }),
+      eventDataPromise,
     };
   }
 
@@ -156,6 +142,11 @@ export default class ContractMethodSender<
     transactionOptions: TransactionOptions,
   ) {
     return this.client.send(this.functionName, callArgs, transactionOptions);
+  }
+
+  async _waitForTransactionReceipt(transactionHash: string) {
+    await this.client.adapter.waitForTransaction(transactionHash);
+    return this.client.adapter.getTransactionReceipt(transactionHash);
   }
 
   /**